+++ src/components/ThemesProduct/ThemesProductItem.tsx
@@ -3,8 +3,8 @@ import {
   StyledThemesProductGridContainer,
   StyledThemesProductPaddingContainer,
 } from './StyledThemesProductItem';
-import { usePresentThemeFetch } from './useThemesProductItem';
-import { useIntersectionObserver } from './useIntersectionObserver';
+import { usePresentThemeFetch } from '../../hooks/useThemesProductItem';
+import { useIntersectionObserver } from '../../hooks/useIntersectionObserver';
 import PresentProductList from '../Home/PresentRanking/Item/PresentRankingItem';
 import { ThemesProductionLabel } from './ThemesProductionLabel';
